+++ frontend/packs/request_to_guild.jsx
@@ -1,0 +1,4 @@
+import RequestToGuild from "components_react/request_to_guild/request_to_guild";
+import WebpackerReact from "webpacker-react";
+
+WebpackerReact.setup({RequestToGuild});